refactor: create wrapper for vec<DocumentEmbeddings> for from/tryfrom traits
